+++ ksef-client/src/main/java/pl/akmf/ksef/sdk/api/KsefApiProperties.java
@@ -6,6 +6,7 @@ import java.util.Map;
 public abstract class KsefApiProperties {
 
     public abstract String getBaseUri();
+    public abstract String getSuffixUri();
     public abstract String getQrUri();
 
     public abstract Duration getRequestTimeout();